chore: refactored ObjectIdUtils
